separate theming config
